+++ quilt-sync/ui/src/components.rs
@@ -1,12 +1,12 @@
 pub mod buttons;
 pub mod ignore_popup;
 pub mod layout;
-pub mod set_origin_popup;
+pub mod set_remote_popup;
 pub mod spinner;
 pub mod update_checker;
 
 pub use ignore_popup::{IgnorePopup, IgnorePopupData, UnignorePopup, UnignorePopupData};
 pub use layout::{Layout, Notification, ToolbarActions};
-pub use set_origin_popup::{SetOriginPopup, SetOriginPopupData};
+pub use set_remote_popup::{SetRemotePopup, SetRemotePopupData};
 pub use spinner::Spinner;
 pub use update_checker::UpdateChecker;